+++ prefixmap/src/main/java/nl/basjes/collections/PrefixMap.java
@@ -18,6 +18,7 @@ package nl.basjes.collections;
 
 import java.io.Serializable;
 import java.util.Collection;
+import java.util.Iterator;
 import java.util.Map;
 import java.util.Set;
 
@@ -224,6 +225,18 @@ public interface PrefixMap<V extends Serializable> extends Serializable, Map<Str
     V getLongestMatch(String input);
 
     /**
+     * <p>Returns List of all matches that have a value.</p>
+     * <p>The list contains all non-null values for the prefix values where this is true:
+     * <code>input.startsWith(prefix)</code>.</p>
+     * <p>Note that implementations may be constructed to match either
+     * case sensitive or case insensitive.</p>
+     *
+     * @param input The string for which we need value of the stored prefix
+     * @return The list of values, an empty List if nothing is found.
+     */
+    Iterator<V> getAllMatches(String input);
+
+    /**
      * <p>Return the value of the <code>shortest</code> matching prefix. </p>
      * <p>The value returned is the shortest stored prefix for which is true:
      * <code>input.startsWith(prefix)</code>.</p>
@@ -246,4 +259,17 @@ public interface PrefixMap<V extends Serializable> extends Serializable, Map<Str
      * @return The value, null if not found.
      */
     V getLongestMatch(char[] input);
+
+    /**
+     * <p>Returns List of all matches that have a value.</p>
+     * <p>The list contains all non-null values for the prefix values where this is true:
+     * <code>input.startsWith(prefix)</code>.</p>
+     * <p>Note that implementations may be constructed to match either
+     * case sensitive or case insensitive.</p>
+     *
+     * @param input The string for which we need value of the stored prefix
+     * @return The list of values, an empty List if nothing is found.
+     */
+    Iterator<V> getAllMatches(char[] input);
+
 }