+++ prefixmap/src/main/java/nl/basjes/collections/prefixmap/MatchFunctionString.java
@@ -1,5 +1,0 @@
-package nl.basjes.collections.prefixmap;
-
-public interface MatchFunction<R, V> {
-    apply
-}