+++ prefixmap/src/test/java/nl/basjes/collections/prefixmap/TestAllMapMethods.java
@@ -19,6 +19,8 @@ package nl.basjes.collections.prefixmap;
 import nl.basjes.collections.PrefixMap;
 import org.junit.jupiter.api.Test;
 
+import java.util.Collections;
+import java.util.Iterator;
 import java.util.Set;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -73,7 +75,7 @@ class TestAllMapMethods {
 
             @Override
             public Set<Entry<String, String>> entrySet() {
-                return null;
+                return Collections.emptySet();
             }
 
             @Override
@@ -97,6 +99,11 @@ class TestAllMapMethods {
             }
 
             @Override
+            public Iterator<String> getAllMatches(String input) {
+                return Collections.emptyIterator();
+            }
+
+            @Override
             public String getShortestMatch(char[] input) {
                 return null;
             }
@@ -104,6 +111,11 @@ class TestAllMapMethods {
             @Override
             public String getLongestMatch(char[] input) {
                 return null;
+            }
+
+            @Override
+            public Iterator<String> getAllMatches(char[] input) {
+                return Collections.emptyIterator();
             }
         };